feat: fix audio ranges calculations
