+++ index.html
@@ -78,7 +78,7 @@
     <div>
       <video id="video" autoplay="true" class="picCapture"></video>
       <video id="someone" controls autoplay muted src="https://upload.wikimedia.org/wikipedia/commons/transcoded/c/c4/Physicsworks.ogv/Physicsworks.ogv.240p.vp9.webm"></video>
-
+<!-- <div id="gradient"></div> -->
       <!-- <video style="display: none;" id="someone" controls autoplay muted loop>
         <source src="./src/assets/videoplayback.mp4" type="video/mp4" />
       </video> -->
