+++ src/javascript/utils/setup-canvas.js
@@ -1,104 +1,14 @@
-import { startAudioVisual } from "../drawings/multiCanvas.js";
+import handleMicrophone from "./microphone.js";
+import { connectMidi } from "./midi-controller.js";
+import connectWebCam from "./webcam.js";
 import { addCanvas, createButtons } from "../helpers/drawSettings.js";
-import CanvasRecorder from "../utils/recorder.js";
-import { getIndexFromValue, getPercentage } from "../helpers/math.js";
+import { CanvasRecorder } from "../utils/recorder.js";
+import { handleRecording } from "../utils/recorder.js";
+import { settings } from "./layer-settings.js";
+
 let size = 1,
-  listening = false,
   WIDTH = window.innerWidth,
   HEIGHT = window.innerHeight;
-
-export const settings = {
-  range: {
-    icon: "fa-assistive-listening-systems",
-    list: ["low bass", "bass", "tenor", "alto", "soprano", "all"],
-    value: "all",
-  },
-  pattern: {
-    icon: "fa-route",
-    list: [
-      "center",
-      "line",
-      "spiral",
-      "diagonal",
-      "grid",
-      "wave",
-      "verticalWave",
-      "circle",
-      "cursor",
-      "random",
-      "clock",
-    ],
-    value: "line",
-  },
-  shape: {
-    icon: "fa-shapes",
-    list: ["triangle", "square", "circle", "star", "ninja"],
-    value: "square",
-  },
-  size: {
-    icon: "fa-search-plus",
-    min: 0,
-    max: 15,
-    value: 7,
-  },
-  stroke: {
-    icon: "fa-pen",
-    checked: true,
-  },
-  color: {
-    icon: "fa-palette",
-    value: "#ff0f22",
-  },
-  opacity: {
-    icon: "fa-eye",
-    min: 0,
-    max: 100,
-    value: 70,
-  },
-  effect: {
-    icon: "fa-chess-board",
-    list: [
-      "source-over",
-      "multiply",
-      "lighten",
-      "difference",
-      "exclusion",
-      "color-dodge",
-      "luminosity",
-      "darken",
-      "screen",
-      "overlay",
-      "xor",
-      "copy",
-      "destination-atop",
-      "destination-over",
-      "destination-out",
-      "destination-in",
-      "source-out",
-      "source-in",
-      "source-atop",
-    ],
-    value: "source-over",
-  },
-  twist: {
-    icon: "fa-biohazard",
-    checked: true,
-  },
-  rotationSpeed: {
-    icon: "fa-sync",
-    min: -10,
-    max: 10,
-    value: 1,
-  },
-};
-
-function hexToRGB(hexColor) {
-  return {
-    r: (hexColor >> 16) & 0xff,
-    g: (hexColor >> 8) & 0xff,
-    b: hexColor & 0xff,
-  };
-}
 
 function appendImage(canvas, target, downloadButton) {
   // set canvasImg image src to data
@@ -107,56 +17,6 @@ function appendImage(canvas, target, downloadButton) {
     downloadButton.href = URL.createObjectURL(blob);
     downloadButton.download = `${Date.now()}-audiovisual.png`;
   }, "image/png");
-}
-
-function handleMicrophone(button, main, controlBoard, settings) {
-  if (button.classList.contains("controller__button-start")) {
-    listening = true;
-    button.style.color = "red";
-    button.classList.remove("controller__button-start");
-    button.classList.add("controller__button-pause");
-    button.innerHTML = "Listening  <i class='fa fa-volume-up'></i>";
-    button.classList.toggle("blink", listening);
-    startAudioVisual(main, controlBoard, settings);
-    document.querySelector("#someone").className = ".hidden";
-  } else {
-    listening = false;
-    button.style.color = "#ccc";
-    button.classList.remove("controller__button-pause");
-    button.classList.add("controller__button-start");
-    button.innerHTML = "Start  <i class='fa fa-play'></i>";
-    button.classList.toggle("blink", listening);
-    document.querySelector("#someone").className = "";
-  }
-}
-
-function handleRecording(button, recorder) {
-  if (button.classList.contains("controller__button-download")) {
-    recorder.save("canvas-recording");
-    button.classList.remove("controller__button-download");
-    button.classList.add("controller__button-record");
-    button.innerHTML = "<i class='fa fa-circle'></i> Record";
-    button.style.color = "#cccccc";
-    return;
-  }
-
-  if (button.classList.contains("controller__button-stop")) {
-    recorder.stop();
-    listening = false;
-    button.classList.toggle("blink");
-    button.classList.remove("controller__button-stop");
-    button.classList.add("controller__button-download");
-    button.innerHTML = "<i class='fa fa-download'></i> Download";
-  }
-
-  if (button.classList.contains("controller__button-record")) {
-    recorder.start();
-    button.classList.toggle("blink");
-    button.classList.remove("controller__button-record");
-    button.classList.add("controller__button-stop");
-    button.innerHTML = "<i class='fa fa-stop-circle'></i> Stop Record";
-    button.style.color = "Red";
-  }
 }
 
 function loadDrawingFromParams(parent, settings) {
@@ -174,102 +34,6 @@ function loadDrawingFromParams(parent, settings) {
     console.log("urlSettings", settings);
     createButtons(parent, settings, index);
     index++;
-  }
-}
-
-function connectWebCam() {
-  var video = document.querySelector("#someone");
-
-  if (navigator.mediaDevices.getUserMedia) {
-    navigator.mediaDevices.getUserMedia({ video: true })
-      .then(function (stream) {
-        video.srcObject = stream;
-      })
-      .catch(function (err0r) {
-        console.log("Something went wrong!");
-      });
-  }
-}
-
-function connectMidi(controls) {
-  navigator.requestMIDIAccess().then(onMIDISuccess, onMIDIFailure);
-
-  function onMIDISuccess(midiAccess) {
-    for (var input of midiAccess.inputs.values()) {
-      input.onmidimessage = getMIDIMessage;
-    }
-  }
-
-  function getMIDIMessage(midiMessage) {
-    const control = midiMessage.data[1];
-    const value = midiMessage.data[2];
-    const colorWell = document.querySelector(".controller__slider-color-1");
-
-    switch (control) {
-      case 3:
-        const rangeControl = document.querySelector(
-          ".controller__select-range-1"
-        );
-        const ranges = settings.range.list;
-        rangeControl.value =
-          ranges[getIndexFromValue(ranges.length - 1, value)];
-        // const size = document.querySelector(".controller__slider-size-1");
-        // size.value = value / 10;
-        break;
-      case 4:
-        const patternControl = document.querySelector(
-          ".controller__select-pattern-1"
-        );
-        const patterns = settings.pattern.list;
-        patternControl.value =
-          patterns[getIndexFromValue(patterns.length - 1, value)];
-        break;
-      case 5:
-        const shape = document.querySelector(".controller__select-shape-1");
-        const shapes = settings.shape.list;
-        shape.value = shapes[getIndexFromValue(shapes.length - 1, value)];
-        break;
-      case 6:
-        // const shape = document.querySelector(".controller__select-shape-1");
-        // const shapes = settings.shape.list;
-        // shape.value = shapes[Math.round(shapes.length * ((1 / 127) * value))];
-        // console.log(shapes.length * ((1 / 127) * value));
-        break;
-      case 14:
-        // rethink all of this, you get hex code but can only submit rgb.
-        // you can use hexToRGB to parse but you need to encode back after
-        const red = 255 * getPercentage(127, value);
-        colorWell.value = {
-          ...colorWell.value,
-          r: red,
-        };
-        console.log({ red, ...colorWell.value });
-        break;
-      case 15:
-        const green = 255 * getPercentage(127, value);
-        colorWell.value = {
-          ...colorWell.value,
-          g: green,
-        };
-
-        break;
-      case 16:
-        const blue = 255 * getPercentage(127, value);
-        colorWell.value = {
-          ...colorWell.value,
-          b: blue,
-        };
-
-        break;
-      default:
-        break;
-    }
-
-    console.log({ control, value });
-  }
-
-  function onMIDIFailure() {
-    console.log("Could not access your MIDI devices.");
   }
 }
 
@@ -329,7 +93,7 @@ window.onload = () => {
   });
 
   shortenButton.addEventListener("click", () => {
-    listening && handleMicrophone(startButton, main, controlBoard, settings);
+    window.listening && handleMicrophone(startButton, main, controlBoard, settings);
     shortenUrl();
   });
   recordButton.addEventListener("click", () =>
@@ -349,4 +113,4 @@ window.onload = () => {
   });
 };
 
-export { size, listening, WIDTH, HEIGHT, appendImage, hexToRGB };
+export { size, WIDTH, HEIGHT, appendImage };
