+++ src/javascript/utils/webcam.js
@@ -1,0 +1,14 @@
+export default function connectWebCam() {
+  var video = document.querySelector("#someone");
+
+  if (navigator.mediaDevices.getUserMedia) {
+    navigator.mediaDevices
+      .getUserMedia({ video: true })
+      .then(function (stream) {
+        video.srcObject = stream;
+      })
+      .catch(function (err0r) {
+        console.log("Something went wrong!");
+      });
+  }
+}
